Pretty Format Response
Summary: This change is meant to improve the format response of the SQL plugin. The protocol for extracting resulting information is meant to be flexible so it can be used for multiple output formats we wish to support. This particular commit only has support for "jdbc" format.

CR: https://code.amazon.com/reviews/CR-3980304

SIM: https://sim.amazon.com/issues/CloudSearch-7704

TEST: Integration tests added and passed.