Pretty Format Response
Summary: This change is meant to improve the format response of the SQL plugin. The protocol for extracting resulting information is meant to be flexible so it can be used for multiple output formats we wish to support. One of the most important things being including schema information for JDBC to use.

CR: https://code.amazon.com/reviews/CR-3980304

SIM: https://sim.amazon.com/issues/CloudSearch-7704

TEST: Integ tests added and passed.