Fixing failing integ tests for mvn build (#49)
Issue #, if available: N/A

Description of changes: After introducing new data in test indexes (see #42 ), some of the integ tests were broken. This change fixes those.

We didn't notice this, since we rely on gradle build mostly, which does not run the mvn integ tests. Once #20 is fixed all the remaining mvn tests can be migrated to the new ESIntegTest subclasses and we can deprecate mvn.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.