Issue #, if available: #44
Description of changes: The order of fields in the query was not considered
when returning the results in csv format. This change fixes that problem.

How tested: tested manually on a local cluster, plus added unit/integ
tests to cover the use case + added code Build with all tests succeeds.