Moving the execution of subqueries from a transport to a worker thread.
Description: If a query has a subquery, the replacement of the subquery
with actual values takes place before the main query execution and on
the same thread that handles the request - the transport thread, which
is not to be used for blocking actions. This change moves the execution
of the subquery to the sql worker threadpool.

How Tested: if '-ea' (enable assertions) jvm options is enabled, any
query with a subquery would fail because of this issue. Verified that
with this change the query succeeds. Sample query used:
'SELECT firstname FROM account WHERE firstname IN (SELECT firstname FROM
people)'. Also added integ tests for the same query.